+++ packages/lint/src/index.ts
@@ -2,6 +2,7 @@ import abstractRules from './rules/abstract-rules.js';
 import buttonRules from './rules/button-rules.js';
 import captionRules from './rules/caption-rules.js';
 import containerRules from './rules/container-rules.js';
+import formErrorRules from './rules/form-error-rules.js';
 import imageRules from './rules/image-rules.js';
 import kickerRules from './rules/kicker-rules.js';
 import landingLeadRules from './rules/landing-lead-rules.js';
@@ -20,6 +21,7 @@ const rules = {
   ...buttonRules.rules,
   ...captionRules.rules,
   ...containerRules.rules,
+  ...formErrorRules.rules,
   ...imageRules.rules,
   ...kickerRules.rules,
   ...landingLeadRules.rules,
@@ -41,6 +43,7 @@ const configs = {
       ...buttonRules.configs.recommended.rules,
       ...captionRules.configs.recommended.rules,
       ...containerRules.configs.recommended.rules,
+      ...formErrorRules.configs.recommended.rules,
       ...imageRules.configs.recommended.rules,
       ...kickerRules.configs.recommended.rules,
       ...landingLeadRules.configs.recommended.rules,
@@ -72,6 +75,7 @@ export type {
   CaptionTagAnalysis,
   ContainerClassInfo,
   ContainerTagAnalysis,
+  FormErrorTagAnalysis,
   HeadingSubheadingBlockTagAnalysis,
   KickerClassInfo,
   KickerTagAnalysis,